fix: extract interface to common type folder for fixing Docker build
