+++ types/app.d.ts
@@ -1,5 +1,12 @@
-import type { RoadizWebResponse } from '@roadiz/types'
+import type { RoadizWebResponse, RoadizWalker } from '@roadiz/types'
 
 export type PageEntityProps = {
     webResponse?: RoadizWebResponse
+}
+
+export interface UseRoadizBlockProps {
+    walker: RoadizWalker
+    index: number
+    blocks?: RoadizWalker[]
+    numBlocks?: number | string
 }
